Update README.md with spam prevention (#35)
I've added a section on how you can prevent mail to be treated as spam. This can be quite common when you change the `from address` to that of a user email address. In fact, it usually will occur if the `from address` domain is not linked to the sending host